+++ src/main/java/net/sf/jabref/gui/preftabs/FileTab.java
@@ -17,6 +17,7 @@ package net.sf.jabref.gui.preftabs;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
+import java.awt.event.ItemListener;
 import javax.swing.BorderFactory;
 import javax.swing.ButtonGroup;
 import javax.swing.JButton;
@@ -29,11 +30,13 @@ import javax.swing.JSpinner;
 import javax.swing.JTextField;
 import javax.swing.SpinnerNumberModel;
 import net.sf.jabref.Globals;
+import net.sf.jabref.external.ExternalFileTypeEditor;
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.JabRefPreferences;
 
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
+import net.sf.jabref.gui.actions.BrowseAction;
 import net.sf.jabref.gui.help.HelpFiles;
 import net.sf.jabref.gui.help.HelpAction;
 import net.sf.jabref.logic.l10n.Localization;
@@ -62,10 +65,37 @@ class FileTab extends JPanel implements PrefsTab {
     private final JSpinner autoSaveInterval;
     private boolean origAutoSaveSetting;
 
+    private final JTextField fileDir;
+    private final JCheckBox bibLocationAsFileDir;
+    private final JCheckBox bibLocAsPrimaryDir;
+    private final JCheckBox runAutoFileSearch;
+    private final JCheckBox allowFileAutoOpenBrowse;
+    private final JRadioButton useRegExpComboBox;
+    private final JRadioButton matchExactKeyOnly = new JRadioButton(Localization.lang("Autolink only files that match the BibTeX key"));
+    private final JRadioButton matchStartsWithKey = new JRadioButton(Localization.lang("Autolink files with names starting with the BibTeX key"));
+    private final JTextField regExpTextField;
+
 
     public FileTab(JabRefFrame frame, JabRefPreferences prefs) {
         this.prefs = prefs;
         this.frame = frame;
+
+        fileDir = new JTextField(25);
+        bibLocationAsFileDir = new JCheckBox(Localization.lang("Allow file links relative to each bib file's location"));
+        bibLocAsPrimaryDir = new JCheckBox(Localization.lang("Use the bib file location as primary file directory"));
+        bibLocAsPrimaryDir.setToolTipText(Localization.lang("When downloading files, or moving linked files to the "
+                + "file directory, prefer the bib file location rather than the file directory set above"));
+        bibLocationAsFileDir.addChangeListener(e -> bibLocAsPrimaryDir.setEnabled(bibLocationAsFileDir.isSelected()));
+        runAutoFileSearch = new JCheckBox(Localization.lang("When opening file link, search for matching file if no link is defined"));
+        allowFileAutoOpenBrowse = new JCheckBox(Localization.lang("Automatically open browse dialog when creating new file link"));
+        regExpTextField = new JTextField(25);
+        useRegExpComboBox = new JRadioButton(Localization.lang("Use Regular Expression Search"));
+        ItemListener regExpListener = e -> regExpTextField.setEditable(useRegExpComboBox.isSelected());
+        useRegExpComboBox.addItemListener(regExpListener);
+        ButtonGroup buttonGroup = new ButtonGroup();
+        buttonGroup.add(matchExactKeyOnly);
+        buttonGroup.add(matchStartsWithKey);
+        buttonGroup.add(useRegExpComboBox);
 
         openLast = new JCheckBox(Localization.lang("Open last edited databases at startup"));
         backup = new JCheckBox(Localization.lang("Backup old file when saving"));
@@ -127,6 +157,41 @@ class FileTab extends JPanel implements PrefsTab {
         builder.append(reformatFileOnSaveAndExport, 3);
         builder.nextLine();
 
+        builder.appendSeparator(Localization.lang("External file links"));
+        JPanel pan = new JPanel();
+        builder.append(pan);
+
+        lab = new JLabel(Localization.lang("Main file directory") + ':');
+        builder.append(lab);
+        builder.append(fileDir);
+        BrowseAction browse = BrowseAction.buildForDir(this.frame, fileDir);
+        builder.append(new JButton(browse));
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(bibLocationAsFileDir, 3);
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(bibLocAsPrimaryDir, 3);
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(matchStartsWithKey, 3);
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(matchExactKeyOnly, 3);
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(useRegExpComboBox);
+        builder.append(regExpTextField);
+
+        builder.append(new HelpAction(Localization.lang("Help on Regular Expression Search"), HelpFiles.regularExpressionSearchHelp).getHelpButton());
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(runAutoFileSearch, 3);
+        builder.nextLine();
+        builder.append(new JPanel());
+        builder.append(allowFileAutoOpenBrowse);
+        builder.nextLine();
+
         builder.appendSeparator(Localization.lang("Autosave"));
         builder.append(autoSave, 1);
         JButton help = new HelpAction(HelpFiles.autosaveHelp).getHelpButton();
@@ -142,7 +207,7 @@ class FileTab extends JPanel implements PrefsTab {
         builder.append(promptBeforeUsingAutoSave);
         builder.nextLine();
 
-        JPanel pan = builder.getPanel();
+        pan = builder.getPanel();
         pan.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
         setLayout(new BorderLayout());
         add(pan, BorderLayout.CENTER);
@@ -151,6 +216,20 @@ class FileTab extends JPanel implements PrefsTab {
 
     @Override
     public void setValues() {
+        fileDir.setText(prefs.get(Globals.FILE_FIELD + Globals.DIR_SUFFIX));
+        bibLocAsPrimaryDir.setSelected(prefs.getBoolean(JabRefPreferences.BIB_LOC_AS_PRIMARY_DIR));
+        bibLocAsPrimaryDir.setEnabled(bibLocationAsFileDir.isSelected());
+        runAutoFileSearch.setSelected(prefs.getBoolean(JabRefPreferences.RUN_AUTOMATIC_FILE_SEARCH));
+        allowFileAutoOpenBrowse.setSelected(prefs.getBoolean(JabRefPreferences.ALLOW_FILE_AUTO_OPEN_BROWSE));
+        regExpTextField.setText(prefs.get(JabRefPreferences.REG_EXP_SEARCH_EXPRESSION_KEY));
+        if (prefs.getBoolean(JabRefPreferences.AUTOLINK_USE_REG_EXP_SEARCH_KEY)) {
+            useRegExpComboBox.setSelected(true);
+        } else if (prefs.getBoolean(JabRefPreferences.AUTOLINK_EXACT_KEY_ONLY)) {
+            matchExactKeyOnly.setSelected(true);
+        } else {
+            matchStartsWithKey.setSelected(true);
+        }
+
         openLast.setSelected(prefs.getBoolean(JabRefPreferences.OPEN_LAST_EDITED));
         backup.setSelected(prefs.getBoolean(JabRefPreferences.BACKUP));
 
@@ -180,6 +259,16 @@ class FileTab extends JPanel implements PrefsTab {
 
     @Override
     public void storeSettings() {
+        prefs.put(Globals.FILE_FIELD + Globals.DIR_SUFFIX, fileDir.getText());
+        prefs.putBoolean(JabRefPreferences.BIB_LOC_AS_PRIMARY_DIR, bibLocAsPrimaryDir.isSelected());
+        prefs.putBoolean(JabRefPreferences.RUN_AUTOMATIC_FILE_SEARCH, runAutoFileSearch.isSelected());
+        prefs.putBoolean(JabRefPreferences.ALLOW_FILE_AUTO_OPEN_BROWSE, allowFileAutoOpenBrowse.isSelected());
+        prefs.putBoolean(JabRefPreferences.AUTOLINK_USE_REG_EXP_SEARCH_KEY, useRegExpComboBox.isSelected());
+        prefs.putBoolean(JabRefPreferences.AUTOLINK_EXACT_KEY_ONLY, matchExactKeyOnly.isSelected());
+        if (useRegExpComboBox.isSelected()) {
+            prefs.put(JabRefPreferences.REG_EXP_SEARCH_EXPRESSION_KEY, regExpTextField.getText());
+        }
+
         String newline;
         switch (newlineSeparator.getSelectedIndex()) {
         case 0: